+++ tests/fakes/fake_memfault_platform_metrics_locking.h
@@ -10,8 +10,16 @@
 
 #include <stdbool.h>
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 //! @return true if there has been an equivalent number of lock and unlock calls, else false
 bool fake_memfault_platform_metrics_lock_calls_balanced(void);
 
 //! Reset the state of the fake locking tracker
 void fake_memfault_metrics_platorm_locking_reboot(void);
+
+#ifdef __cplusplus
+}
+#endif